configure msw to work with frontend public
