+++ frontend/package.json
@@ -171,17 +171,17 @@
     "@graphql-codegen/typescript": "^4.0.6",
     "@graphql-codegen/typescript-operations": "^4.2.0",
     "@graphql-codegen/typescript-react-apollo": "^4.3.0",
-    "@storybook/addon-a11y": "^8.0.8",
-    "@storybook/addon-actions": "^8.0.8",
-    "@storybook/addon-essentials": "^8.0.8",
-    "@storybook/addon-interactions": "^8.0.8",
-    "@storybook/addon-links": "^8.0.8",
-    "@storybook/blocks": "^8.0.8",
-    "@storybook/cli": "^8.0.8",
+    "@storybook/addon-a11y": "^8.1.1",
+    "@storybook/addon-actions": "^8.1.1",
+    "@storybook/addon-essentials": "^8.1.1",
+    "@storybook/addon-interactions": "^8.1.1",
+    "@storybook/addon-links": "^8.1.1",
+    "@storybook/blocks": "^8.1.1",
+    "@storybook/cli": "^8.1.1",
     "@storybook/node-logger": "^8.0.8",
-    "@storybook/preset-create-react-app": "^8.0.8",
-    "@storybook/react": "^8.0.8",
-    "@storybook/react-webpack5": "^8.0.8",
+    "@storybook/preset-create-react-app": "^8.1.1",
+    "@storybook/react": "^8.1.1",
+    "@storybook/react-webpack5": "^8.1.1",
     "@storybook/testing-library": "^0.2.2",
     "@testing-library/jest-dom": "^6.1.3",
     "@types/apollo-upload-client": "^17.0.0",
@@ -215,16 +215,16 @@
     "jest-fetch-mock": "^3.0.3",
     "jest-location-mock": "^2.0.0",
     "mockdate": "^3.0.5",
-    "msw": "^0.36.4",
-    "msw-storybook-addon": "^1.10.0",
+    "msw": "^2.3.0",
+    "msw-storybook-addon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "otplib": "^12.0.1",
     "prettier": "^2.8.4",
     "redux-mock-store": "^1.5.4",
     "sass": "^1.63.6",
     "selenium-webdriver": "^4.16.0",
-    "storybook": "^8.0.8",
-    "storybook-addon-apollo-client": "^6.0.0",
+    "storybook": "^8.1.1",
+    "storybook-addon-apollo-client": "^7.3.0",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "typescript": "^5.3.2",
@@ -239,7 +239,9 @@
     "webpack@>5.0.0": "5.76.0"
   },
   "msw": {
-    "workerDirectory": "/storybook_public"
+    "workerDirectory": [
+      "public"
+    ]
   },
   "jest": {
     "moduleNameMapper": {
